Move evaluation_metric.py to base subdirectory
