+++ project/modules/performance/metrics/aggregate/hit_rate.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pandas as pd
 
-from ..evaluation_metric import AggregateMetric
+from ..base.evaluation_metric import AggregateMetric
 
 
 class HitRate(AggregateMetric):
