+++ project/modules/performance/metrics/aggregate/standard_deviation_of_return.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pandas as pd
 
-from ..evaluation_metric import AggregateMetric
+from ..base.evaluation_metric import AggregateMetric
 
 
 class StandardDeviationOfReturn(AggregateMetric):
