+++ project/modules/performance/metrics/series/cumulative_return.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pandas as pd
 
-from ..evaluation_metric import SeriesMetric
+from ..base.evaluation_metric import SeriesMetric
 
 
 class CumulativeReturn(SeriesMetric):
